Fix code block immediately following paragraph
in commit 6a996bc4316f02ec315476b295f1d8d078b7fc8d @micbou made
html2vimdoc support preformatted text inside a paragraph, but this was
broken in the specific case where there was no empty line between the
paragraph and the code block.

In that case we actually get a block (list) element containing a
preformatted text block. In that case, recurse into the render for that.